Refactors and updates Confirmation
