+++ lib/hooks/auth/index.ts
@@ -1,3 +1,2 @@
-export { useConfirm } from "./useConfirm";
 export { useLogin } from "./useLogin";
 export { useResetPassword } from "./useResetPassword";
