+++ public/static/locales/en/common.json
@@ -142,5 +142,8 @@
         }
       }
     }
+  },
+  "support": {
+    "needHelp": "Need help?"
   }
 }
