+++ public/static/locales/en/signup.json
@@ -1,8 +1,8 @@
 {
   "signUpConfirmation": {
-    "title": "Check your email",
+    "title": "Check your email for a security code",
     "resendConfirmationCodeButton": "Send a new verification code",
-    "emailHasBeenSent": "An email with a verification code has been sent to",
+    "emailHasBeenSent": "This step helps keep GC Forms secure by verifying your account each time you sign in.",
     "confirmButton": "Continue",
     "resendConfirmationCode": {
       "title": "Check your email",
@@ -10,9 +10,13 @@
     },
     "fields": {
       "confirmationCode": {
-        "label": "Enter the verification code to finish creating your account",
+        "label": "Security code",
+        "description": "Enter the code within 15 minutes.",
         "error": {
-          "number": "This must be a number"
+          "notEmpty": "Cannot be empty",
+          "fiveChars": "Must be 5 characters",
+          "notFound": "Code not found",
+          "expired": "Code expired. Try signing in again."
         }
       }
     }
@@ -53,9 +57,8 @@
   },
   "registrationClosed": "Registration is closed at this time",
   "accountCreated": {
-    "title": "GC Forms account created",
-    "yourAccountTitle": "What you can do with your account",
-    "yourAccountListDescription": "With this account you can",
+    "title": "Account created",
+    "yourAccountTitle": "What can you do",
     "yourAccountList": {
       "item1": "draft a form",
       "item2": "save drafts within GC Forms",
