+++ public/static/locales/fr/common.json
@@ -143,5 +143,8 @@
         }
       }
     }
+  },
+  "support": {
+    "needHelp": "Besoin d’aide?"
   }
 }
